+++ Engine/src/renderers/SpriteRenderer.cpp
@@ -23,6 +23,6 @@ void SpriteRenderer::render(const double rotation)
         this->center = { this->w / 2, this->h / 2 };
     }
 
-    this->textureRect = { this->gameObject->globalPosition.x, this->gameObject->globalPosition.y, this->w, this->h };
+    this->textureRect = { (int)this->gameObject->globalPosition.x, (int)this->gameObject->globalPosition.y, this->w, this->h };
     SDL_RenderCopyEx(this->sdlRenderer, texture, NULL, &this->textureRect, rotation, &this->center, SDL_RendererFlip::SDL_FLIP_NONE);
 }
