+++ wobo/src/game/Wobba.cpp
@@ -9,11 +9,11 @@ Wobba::Wobba() : Game("Wobba", { 800, 600 })
 void Wobba::init()
 {
     auto terrainFactory = TerrainFactory(*this);
-    for (int x = 0; x + 8 <= 800; x += 8) {
+    for (float x = 0; x + 8 <= 800; x += 8) {
         terrainFactory.create({ x, 500 });
     }
 
-    for (int y = 400; y < 500; y += 8) {
+    for (float y = 400; y < 500; y += 8) {
         terrainFactory.create({ 500, y });
     }
